[feat] FirebaseStorageManager 수정
sapmple network 삭제 및 upload 추가, samplefilemanager 삭제
